% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{add_gender_data}
\alias{add_gender_data}
\title{Add gender data to campus enrollment data}
\usage{
add_gender_data(campus_df, gender_df, end_year)
}
\arguments{
\item{campus_df}{Processed campus data frame}

\item{gender_df}{Raw gender data frame}

\item{end_year}{School year end}
}
\value{
Campus data frame with gender columns added
}
\description{
Add gender data to campus enrollment data
}
\keyword{internal}
