% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_sd_excel}
\alias{download_sd_excel}
\title{Download an Excel file from SD DOE}
\usage{
download_sd_excel(url, end_year, skip_rows = 0)
}
\arguments{
\item{url}{Full URL to the Excel file}

\item{end_year}{School year end (for error messages)}

\item{skip_rows}{Number of header rows to skip}
}
\value{
Data frame
}
\description{
Downloads and reads an Excel file from the SD DOE website.
}
\keyword{internal}
