% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for SD enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available years
}
\description{
Returns the range of years available for download from the
South Dakota Department of Education.
}
\examples{
get_available_years()
}
