% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_district_filename}
\alias{get_district_filename}
\title{Get district enrollment filename for a given year}
\usage{
get_district_filename(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Filename string
}
\description{
Get district enrollment filename for a given year
}
\keyword{internal}
