% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_grade_filename}
\alias{get_grade_filename}
\title{Get school-level grade enrollment filename for a given year}
\usage{
get_grade_filename(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Filename string
}
\description{
Get school-level grade enrollment filename for a given year
}
\keyword{internal}
