% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_race_filename}
\alias{get_race_filename}
\title{Get race/ethnicity enrollment filename for a given year}
\usage{
get_race_filename(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Filename string or NULL if not available
}
\description{
Get race/ethnicity enrollment filename for a given year
}
\keyword{internal}
