% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from SD DOE}
\usage{
get_raw_directory()
}
\value{
Combined data frame with school and district directory data
}
\description{
Downloads both the principal/school info file and the superintendent/district
info file from SD DOE and joins them by district number.
}
\keyword{internal}
