% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_era1}
\alias{get_raw_enr_era1}
\title{Download raw enrollment data for Era 1 (2011-2012)}
\usage{
get_raw_enr_era1(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with district and campus data frames
}
\description{
Download raw enrollment data for Era 1 (2011-2012)
}
\keyword{internal}
