% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_era3}
\alias{get_raw_enr_era3}
\title{Download raw enrollment data for Era 3 (2021+)}
\usage{
get_raw_enr_era3(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with district and campus data frames
}
\description{
Download raw enrollment data for Era 3 (2021+)
}
\keyword{internal}
