% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdschooldata-package.R
\docType{package}
\name{sdschooldata-package}
\alias{sdschooldata}
\alias{sdschooldata-package}
\title{sdschooldata: Fetch and Process South Dakota School Data}
\description{
Downloads and processes school data from the South Dakota Department of Education (SD DOE). Provides functions for fetching enrollment data from Fall Census counts and transforming it into tidy format for analysis. Supports historical data from 2006 to present.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almartin82/sdschooldata}
  \item Report bugs at \url{https://github.com/almartin82/sdschooldata/issues}
}

}
\author{
\strong{Maintainer}: Al Martin \email{almartin@example.com}

}
\keyword{internal}
