% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdschooldata-package.R
\docType{package}
\name{sdschooldata-package}
\alias{sdschooldata}
\alias{sdschooldata-package}
\title{sdschooldata: South Dakota School Data}
\description{
A simple, consistent interface for accessing South Dakota school data in Python and R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://almartin82.github.io/sdschooldata}
  \item \url{https://github.com/almartin82/sdschooldata}
  \item Report bugs at \url{https://github.com/almartin82/sdschooldata/issues}
}

}
\author{
\strong{Maintainer}: Al Martin \email{almartin@example.com}

}
\keyword{internal}
